Rename MMTClient -> LibraryClient
