+++ src/Clients/LibraryClient/index.ts
@@ -1,12 +1,12 @@
 
-import MMTClient from "./MMTClient";
+import LibraryClient from "./LibraryClient";
 // tslint:disable-next-line:export-name
-export { default as MMTClient } from "./MMTClient";
+export { default as LibraryClient } from "./LibraryClient";
 
 import MockClient from "./MockClient";
 import RestClient from "./RestClient";
 
 // creates a new client or mock client
-export default function createMMTClient(url: string): MMTClient {
+export default function createLibraryClient(url: string): LibraryClient {
     return url !== "" ? new RestClient(url) : new MockClient();
 }
